Added pixel info in header, and rotate classes.
There is now a rotate-clockwise and rotate-cclockwise class id for
page elements. The page header now shows information about what pixel
is in the crosshairs. I added "VEGAS Data Display" text on the left
side of the page. Also, the spectrum plot y-axis is labeled with a
page element because at the moment I don't know how to make it appear
using dojo.