Moved UI code into this repository.
Originally the repositories were seperate to keep the server
and client code independent.  However, they really needed to
keep in sync and often a change in one related to a change in
the other.  The client code is still largely independent and
resides comletely in the static/ directory.  This should make
the code managment easier.